+++ sdk/resourcemanager/keyvault/armkeyvault/operations_client_example_test.go
@@ -11,7 +11,7 @@ import (
 	"log"
 )
 
-// Generated from example definition: 2026-02-01/listOperations.json
+// Generated from example definition: 2026-03-01-preview/listOperations.json
 func ExampleOperationsClient_NewListPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {